improve structure and reusability
