+++ src/index.html
@@ -33,8 +33,8 @@
         <input
           type="text"
           autofocus
-          class="log-in-name"
-          placeholder="Username"
+          class="log-in-email"
+          placeholder="Your email"
         />
         <input type="password" class="log-in-password" placeholder="Password" />
         <button type="submit" class="submit-btn log-in-btn">Confirm</button>
@@ -44,7 +44,7 @@
       <form class="form sign-up-form hidden-element">
         <h2>Sign Up</h2>
         <p class="sign-up-message"></p>
-        <input type="text" class="sign-up-name" placeholder="Username" />
+        <input type="text" class="sign-up-email" placeholder="Your email" />
         <input
           type="password"
           class="sign-up-password"
